Split rbe510.hpp into separate hpp and cpp files
Conflicts:
	src/rbe510.hpp
