+++ CMakeLists.txt
@@ -21,9 +21,9 @@ find_package(OpenCV REQUIRED)
 include_directories(${OpenCV_INCLUDE_DIRS})
 link_directories(${OpenCV_LIBRARY})
 
-add_executable(rbe_server ${PROJECT_SOURCE_DIR}/src/rbe_server.cpp)
-add_executable(rbe_test_client ${PROJECT_SOURCE_DIR}/src/rbe_test_client.cpp)
-add_executable(g2p ${PROJECT_SOURCE_DIR}/src/g2p.cpp)
+add_executable(rbe_server ${PROJECT_SOURCE_DIR}/src/rbe_server.cpp ${PROJECT_SOURCE_DIR}/src/rbe510.cpp)
+add_executable(rbe_test_client ${PROJECT_SOURCE_DIR}/src/rbe_test_client.cpp ${PROJECT_SOURCE_DIR}/src/rbe510.cpp)
+add_executable(g2p ${PROJECT_SOURCE_DIR}/src/g2p.cpp ${PROJECT_SOURCE_DIR}/src/rbe510.cpp)
 
 target_link_libraries (rbe_server ${OpenCV_LIBS})
 target_link_libraries (rbe_server ${CMAKE_THREAD_LIBS_INIT})
